[script] Reduce log spam and only update lux value if changed for lightsensor

+++ stage2/03-crankshaft-base/files/opt/crankshaft/service_lightsensor.py
@@ -42,6 +42,8 @@ i2cBus = smbus.SMBus(BUS)
 # (scale factor 1)
 i2cBus.write_byte_data(TSL2561_ADDR, 0x80, 0x03)
 
+lastvalue = 0
+
 while True:
   # read global brightness
   # read low byte
@@ -82,28 +84,31 @@ while True:
       Lux = 0.00146*float(Ambient) - 0.00112*float(Infrared)
     else:
       Lux = 0
-    Luxrounded=round(Lux,0)
-    print ("Lux = {}\n".format(Luxrounded))
+    Luxrounded=round(Lux,1)
+    if lastvalue != Luxrounded:
+        print ("Lux = {}\n".format(Luxrounded))
+        os.system("echo {} > /tmp/tsl2561".format(Luxrounded))
+        lastvalue = Luxrounded
 
-    #Set display brigthness
-    if Luxrounded <= level_1:
-        os.system("crankshaft brightness set " + str(display_brigthness_1))
-        if daynight_gpio == 0:
-            os.system("touch /tmp/night_mode_enabled >/dev/null 2>&1")
-    elif Luxrounded > level_1 and Luxrounded < level_2:
-        os.system("crankshaft brightness set " + str(display_brigthness_2))
-        if daynight_gpio == 0:
-            os.system("touch /tmp/night_mode_enabled >/dev/null 2>&1")
-    elif Luxrounded >= level_2 and Luxrounded < level_3:
-        os.system("crankshaft brightness set " + str(display_brigthness_3))
-        if daynight_gpio == 0:
-            os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
-    elif Luxrounded >= level_3 and Luxrounded < level_4:
-        os.system("crankshaft brightness set " + str(display_brigthness_4))
-        if daynight_gpio == 0:
-            os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
-    elif Luxrounded >= level_5:
-        os.system("crankshaft brightness set " + str(display_brigthness_5))
-        if daynight_gpio == 0:
-            os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
+        #Set display brigthness
+        if Luxrounded <= level_1:
+            os.system("crankshaft brightness set " + str(display_brigthness_1))
+            if daynight_gpio == 0:
+                os.system("touch /tmp/night_mode_enabled >/dev/null 2>&1")
+        elif Luxrounded > level_1 and Luxrounded < level_2:
+            os.system("crankshaft brightness set " + str(display_brigthness_2))
+            if daynight_gpio == 0:
+                os.system("touch /tmp/night_mode_enabled >/dev/null 2>&1")
+        elif Luxrounded >= level_2 and Luxrounded < level_3:
+            os.system("crankshaft brightness set " + str(display_brigthness_3))
+            if daynight_gpio == 0:
+                os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
+        elif Luxrounded >= level_3 and Luxrounded < level_4:
+            os.system("crankshaft brightness set " + str(display_brigthness_4))
+            if daynight_gpio == 0:
+                os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
+        elif Luxrounded >= level_5:
+            os.system("crankshaft brightness set " + str(display_brigthness_5))
+            if daynight_gpio == 0:
+                os.system("sudo rm /tmp/night_mode_enabled >/dev/null 2>&1")
   sleep (10)
